+++ sdk/resourcemanager/network/armnetwork/constants.go
@@ -10,7 +10,7 @@ package armnetwork
 
 const (
 	moduleName    = "armnetwork"
-	moduleVersion = "v4.2.0"
+	moduleVersion = "v4.3.0"
 )
 
 // Access - Access to be allowed or denied.
@@ -812,14 +812,16 @@ func PossibleBastionConnectProtocolValues() []BastionConnectProtocol {
 type BastionHostSKUName string
 
 const (
-	BastionHostSKUNameBasic    BastionHostSKUName = "Basic"
-	BastionHostSKUNameStandard BastionHostSKUName = "Standard"
+	BastionHostSKUNameBasic     BastionHostSKUName = "Basic"
+	BastionHostSKUNameDeveloper BastionHostSKUName = "Developer"
+	BastionHostSKUNameStandard  BastionHostSKUName = "Standard"
 )
 
 // PossibleBastionHostSKUNameValues returns the possible values for the BastionHostSKUName const type.
 func PossibleBastionHostSKUNameValues() []BastionHostSKUName {
 	return []BastionHostSKUName{
 		BastionHostSKUNameBasic,
+		BastionHostSKUNameDeveloper,
 		BastionHostSKUNameStandard,
 	}
 }
@@ -1619,20 +1621,24 @@ func PossibleFirewallPolicyIDPSQuerySortOrderValues() []FirewallPolicyIDPSQueryS
 	}
 }
 
-// FirewallPolicyIDPSSignatureDirection - Describes in which direction signature is being enforced: 0 - Inbound, 1 - OutBound,
-// 2 - Bidirectional
+// FirewallPolicyIDPSSignatureDirection - Describes in which direction signature is being enforced: 0 - OutBound, 1 - InBound,
+// 2 - Any, 3 - Internal, 4 - InternalOutbound
 type FirewallPolicyIDPSSignatureDirection int32
 
 const (
-	FirewallPolicyIDPSSignatureDirectionOne  FirewallPolicyIDPSSignatureDirection = 1
-	FirewallPolicyIDPSSignatureDirectionTwo  FirewallPolicyIDPSSignatureDirection = 2
-	FirewallPolicyIDPSSignatureDirectionZero FirewallPolicyIDPSSignatureDirection = 0
+	FirewallPolicyIDPSSignatureDirectionFour  FirewallPolicyIDPSSignatureDirection = 4
+	FirewallPolicyIDPSSignatureDirectionOne   FirewallPolicyIDPSSignatureDirection = 1
+	FirewallPolicyIDPSSignatureDirectionThree FirewallPolicyIDPSSignatureDirection = 3
+	FirewallPolicyIDPSSignatureDirectionTwo   FirewallPolicyIDPSSignatureDirection = 2
+	FirewallPolicyIDPSSignatureDirectionZero  FirewallPolicyIDPSSignatureDirection = 0
 )
 
 // PossibleFirewallPolicyIDPSSignatureDirectionValues returns the possible values for the FirewallPolicyIDPSSignatureDirection const type.
 func PossibleFirewallPolicyIDPSSignatureDirectionValues() []FirewallPolicyIDPSSignatureDirection {
 	return []FirewallPolicyIDPSSignatureDirection{
+		FirewallPolicyIDPSSignatureDirectionFour,
 		FirewallPolicyIDPSSignatureDirectionOne,
+		FirewallPolicyIDPSSignatureDirectionThree,
 		FirewallPolicyIDPSSignatureDirectionTwo,
 		FirewallPolicyIDPSSignatureDirectionZero,
 	}
@@ -1656,7 +1662,7 @@ func PossibleFirewallPolicyIDPSSignatureModeValues() []FirewallPolicyIDPSSignatu
 	}
 }
 
-// FirewallPolicyIDPSSignatureSeverity - Describes the severity of signature: 1 - Low, 2 - Medium, 3 - High
+// FirewallPolicyIDPSSignatureSeverity - Describes the severity of signature: 1 - High, 2 - Medium, 3 - Low
 type FirewallPolicyIDPSSignatureSeverity int32
 
 const (
@@ -1671,6 +1677,26 @@ func PossibleFirewallPolicyIDPSSignatureSeverityValues() []FirewallPolicyIDPSSig
 		FirewallPolicyIDPSSignatureSeverityOne,
 		FirewallPolicyIDPSSignatureSeverityThree,
 		FirewallPolicyIDPSSignatureSeverityTwo,
+	}
+}
+
+// FirewallPolicyIntrusionDetectionProfileType - Possible Intrusion Detection profile values.
+type FirewallPolicyIntrusionDetectionProfileType string
+
+const (
+	FirewallPolicyIntrusionDetectionProfileTypeAdvanced FirewallPolicyIntrusionDetectionProfileType = "Advanced"
+	FirewallPolicyIntrusionDetectionProfileTypeBasic    FirewallPolicyIntrusionDetectionProfileType = "Basic"
+	FirewallPolicyIntrusionDetectionProfileTypeExtended FirewallPolicyIntrusionDetectionProfileType = "Extended"
+	FirewallPolicyIntrusionDetectionProfileTypeStandard FirewallPolicyIntrusionDetectionProfileType = "Standard"
+)
+
+// PossibleFirewallPolicyIntrusionDetectionProfileTypeValues returns the possible values for the FirewallPolicyIntrusionDetectionProfileType const type.
+func PossibleFirewallPolicyIntrusionDetectionProfileTypeValues() []FirewallPolicyIntrusionDetectionProfileType {
+	return []FirewallPolicyIntrusionDetectionProfileType{
+		FirewallPolicyIntrusionDetectionProfileTypeAdvanced,
+		FirewallPolicyIntrusionDetectionProfileTypeBasic,
+		FirewallPolicyIntrusionDetectionProfileTypeExtended,
+		FirewallPolicyIntrusionDetectionProfileTypeStandard,
 	}
 }